Extract has_no_code tests + Create the rest of has_no_code tests
