+++ examples/safe-erc20/tests/erc20.rs
@@ -53,33 +53,6 @@ async fn safe_transfers(alice: Account, bob: Account) -> eyre::Result<()> {
 }
 
 #[e2e::test]
-async fn safe_transfer_rejects_with_eoa_as_token(
-    alice: Account,
-    bob: Account,
-    has_no_code: Account,
-) -> eyre::Result<()> {
-    let safe_erc20_mock_addr = alice.as_deployer().deploy().await?.address()?;
-    let safe_erc20_mock_alice =
-        SafeErc20::new(safe_erc20_mock_addr, &alice.wallet);
-    let bob_addr = bob.address();
-    let has_no_code_addr = has_no_code.address();
-
-    let value = uint!(1_U256);
-
-    let err = send!(safe_erc20_mock_alice.safeTransfer(
-        has_no_code_addr,
-        bob_addr,
-        value
-    ))
-    .expect_err("should not be able to invoke 'transfer' on EOA");
-    assert!(err.reverted_with(SafeErc20::SafeErc20FailedOperation {
-        token: has_no_code_addr
-    }));
-
-    Ok(())
-}
-
-#[e2e::test]
 async fn safe_transfer_rejects_on_error(
     alice: Account,
     bob: Account,
@@ -144,35 +117,6 @@ async fn safe_transfers_from(alice: Account, bob: Account) -> eyre::Result<()> {
 
     assert_eq!(initial_alice_balance - value, alice_balance);
     assert_eq!(initial_bob_balance + value, bob_balance);
-
-    Ok(())
-}
-
-#[e2e::test]
-async fn safe_transfer_from_rejects_with_eoa_as_token(
-    alice: Account,
-    bob: Account,
-    has_no_code: Account,
-) -> eyre::Result<()> {
-    let safe_erc20_mock_addr = alice.as_deployer().deploy().await?.address()?;
-    let safe_erc20_mock_alice =
-        SafeErc20::new(safe_erc20_mock_addr, &alice.wallet);
-    let alice_addr = alice.address();
-    let bob_addr = bob.address();
-    let has_no_code_addr = has_no_code.address();
-
-    let value = uint!(1_U256);
-
-    let err = send!(safe_erc20_mock_alice.safeTransferFrom(
-        has_no_code_addr,
-        alice_addr,
-        bob_addr,
-        value
-    ))
-    .expect_err("should not be able to invoke 'transferFrom' on EOA");
-    assert!(err.reverted_with(SafeErc20::SafeErc20FailedOperation {
-        token: has_no_code_addr
-    }));
 
     Ok(())
 }
